requirements.txt
fastapi
uvicorn
streamlit
langgraph
ollama
faiss-cpu
duckduckgo-search
numpy